# swagger.json
{"openapi":"3.0.0","paths":{"/api/login":{"post":{"operationId":"AuthController_login","summary":"","description":"Log in via email and password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}}},"tags":["auth"]}},"/api/me":{"get":{"operationId":"AuthController_getme","summary":"","description":"Get user entity","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OmitTypeClass"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/api/refresh":{"post":{"operationId":"AuthController_refresh","summary":"","description":"Refresh auth token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}}},"tags":["auth"]}},"/api/students":{"post":{"operationId":"UsersController_addUsers","summary":"","description":"Add students\nRoles: ORGANIZATOR","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddStudentsDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["students"]}},"/api/teachers":{"post":{"operationId":"UsersController_addTeachers","summary":"","description":"Add teachers\nRoles: ORGANIZATOR","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddTeachersDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["teachers"]}},"/api/students/{id}":{"delete":{"operationId":"UsersController_removeStudent","summary":"","description":"Remove student\nRoles: ORGANIZATOR","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["students"]}},"/api/students/courses":{"get":{"operationId":"UsersController_getCourses","summary":"","description":"Get all student courses\nRoles: STUDENT","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseModel"}}}}}},"security":[{"bearer":[]}],"tags":["students"]}},"/metrics":{"get":{"operationId":"PrometheusController_index","parameters":[],"responses":{"200":{"description":""}}}},"/api/tests/{id}/questions":{"post":{"operationId":"TestsController_addQuestion","summary":"","description":"Add new question to test","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddQuestionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tests"],"security":[{"bearer":[]}]}},"/api/tests/{id}/start-session":{"post":{"operationId":"TestsController_startSession","summary":"","description":"Start test session","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["tests"],"security":[{"bearer":[]}]}},"/api/tests/session/{id}":{"post":{"operationId":"TestsController_addAnswer","summary":"","description":"Add answer within the session\nRoles: STUDENT","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddAnswerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tests"],"security":[{"bearer":[]}]}},"/api/tests/session/{id}/finish":{"post":{"operationId":"TestsController_saveSession","summary":"","description":"Save session and get results\nRoles: STUDENT","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["tests"],"security":[{"bearer":[]}]}},"/api/tests/{id}/results":{"get":{"operationId":"TestsController_getResults","summary":"","description":"Get all test results\nRoles: TEACHER","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["tests"],"security":[{"bearer":[]}]}},"/api/groups":{"get":{"operationId":"GroupsController_getAll","summary":"","description":"Get all organization groups\nRoles: TEACHER, ORGANIZATOR","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Group"}}}}}},"tags":["groups"],"security":[{"bearer":[]}]},"post":{"operationId":"GroupsController_createGroup","summary":"","description":"Create group\nRoles: ORGANIZATOR","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/students":{"get":{"operationId":"GroupsController_getStudents","summary":"","description":"Get all group students\nRoles: TEACHER, ORGANIZATOR","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}":{"put":{"operationId":"GroupsController_updateGroup","summary":"","description":"Update group\nRoles: ORGANIZATOR","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroupDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/courses/{id}/lessons":{"get":{"operationId":"CoursesController_getLessons","summary":"","description":"Get course lessons","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["courses"],"security":[{"bearer":[]}]}},"/api/courses/{id}":{"get":{"operationId":"CoursesController_findOne","summary":"","description":"Get course by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["courses"],"security":[{"bearer":[]}]}},"/api/courses":{"post":{"operationId":"CoursesController_createCourse","summary":"","description":"Create course\nRoles: TEACHER","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCourseDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["courses"],"security":[{"bearer":[]}]}},"/api/courses/{id}/groups":{"get":{"operationId":"CoursesController_getGroups","summary":"","description":"Get course groups\nRoles: TEACHER","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["courses"],"security":[{"bearer":[]}]}},"/api/subjects":{"get":{"operationId":"SubjectsController_getAll","summary":"","description":"Get all organization subjects\nRoles: ORGANIZATOR","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Subject"}}}}}},"tags":["subjects"],"security":[{"bearer":[]}]},"post":{"operationId":"SubjectsController_create","summary":"","description":"Create organization\nRoles: ORGANIZATOR","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubjectDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}}}},"tags":["subjects"],"security":[{"bearer":[]}]}},"/api/subjects/{id}":{"delete":{"operationId":"SubjectsController_delete","summary":"","description":"Delete organization\nRoles: ORGANIZATOR","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}}}},"tags":["subjects"],"security":[{"bearer":[]}]}},"/api/lessons/{id}":{"get":{"operationId":"LessonsController_getLesson","summary":"","description":"Get lesson by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonModel"}}}}},"tags":["lessons"],"security":[{"bearer":[]}]}},"/api/course/{id}/lessons":{"post":{"operationId":"LessonsController_createLesson","summary":"","description":"Create lesson\nRoles: TEACHER","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateLessonDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["lessons","courses"],"security":[{"bearer":[]}]}},"/api/teacher/courses":{"get":{"operationId":"TeachersController_getCourses","summary":"","description":"Get all teacher courses\nRoles: TEACHER","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseModel"}}}}}},"tags":["teachers"],"security":[{"bearer":[]}]}}},"info":{"title":"Nestjs FTW","description":"The nestjs API description","version":"1.5","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":8}},"required":["email","password"]},"Token":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"UserGroup":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["id","name"]},"OmitTypeClass":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"role":{"type":"string","enum":["ADMIN","ORGANIZATOR","STUDENT","TEACHER"]},"email":{"type":"string"},"group":{"$ref":"#/components/schemas/UserGroup"}},"required":["id","name","role","email","group"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"AddStudentDto":{"type":"object","properties":{"name":{"type":"string","description":"Student name"},"email":{"type":"string","description":"Student work email"},"groupId":{"type":"number","description":"Group id"}},"required":["name","email","groupId"]},"AddStudentsDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AddStudentDto"}}},"required":["data"]},"AddTeacherDto":{"type":"object","properties":{"name":{"type":"string","description":"Student name"},"email":{"type":"string","description":"Student work email"}},"required":["name","email"]},"AddTeachersDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AddTeacherDto"}}},"required":["data"]},"LessonModel":{"type":"object","properties":{"id":{"type":"number"},"course_id":{"type":"number"},"title":{"type":"string"},"presentation_url":{"type":"string"},"lecture_url":{"type":"string"}},"required":["id","course_id","title","presentation_url","lecture_url"]},"CourseModel":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"subject_id":{"type":"number"},"teacher_id":{"type":"number"},"lessons":{"type":"array","items":{"$ref":"#/components/schemas/LessonModel"}}},"required":["id","title","subject_id","teacher_id","lessons"]},"AnswerEntity":{"type":"object","properties":{"text":{"type":"string"},"correct":{"type":"boolean"}},"required":["text","correct"]},"AddQuestionDto":{"type":"object","properties":{"text":{"type":"string"},"answers":{"type":"array","items":{"$ref":"#/components/schemas/AnswerEntity","maxLength":8,"minLength":1}},"type":{"type":"string","enum":["SINGLE","MULTI"]}},"required":["text","answers","type"]},"AddAnswerDto":{"type":"object","properties":{"questionId":{"type":"number"},"answerId":{"type":"number"}},"required":["questionId","answerId"]},"Group":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"organization_id":{"type":"number"}},"required":["id","name","organization_id"]},"User":{"type":"object","properties":{}},"UpdateGroupDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateGroupDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateCourseDto":{"type":"object","properties":{"title":{"type":"string"},"subjectId":{"type":"number"},"groupIds":{"type":"array","items":{"type":"number"}},"icon":{"type":"string","minLength":1,"maxLength":1}},"required":["title","subjectId","groupIds","icon"]},"Subject":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"organization_id":{"type":"number"}},"required":["id","name","organization_id"]},"CreateSubjectDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateLessonDto":{"type":"object","properties":{"title":{"type":"string"},"presentation":{"type":"string","format":"binary"},"lecture":{"type":"string","format":"binary"}},"required":["title","presentation","lecture"]}}}}